requirement.txt
asgiref==3.2.8
beautifulsoup4==4.9.1
Django==3.0.7
django-bootstrap4==2.0.1
django-crispy-forms==1.9.1
django-webpack-loader==0.7.0
djangorestframework==3.11.0
importlib-metadata==1.6.1
pytz==2020.1
soupsieve==2.0.1
sqlparse==0.3.1
zipp==3.1.0
